Do not cast integer literal as u32
Clippy emits:

  warning: casting integer literal to `u32` is unnecessary

Remove the casts as suggested.

+++ src/cryptographic_primitives/secret_sharing/feldman_vss.rs
@@ -150,7 +150,7 @@ where
         let points = indices
             .iter()
             .map(|i| {
-                let index_bn = BigInt::from(*i as u32 + 1 as u32);
+                let index_bn = BigInt::from(*i as u32 + 1);
                 ECScalar::from(&index_bn)
             })
             .collect::<Vec<P::Scalar>>();
@@ -241,7 +241,7 @@ where
         // add one to indices to get points
         let points: Vec<P::Scalar> = (0..params.share_count)
             .map(|i| {
-                let index_bn = BigInt::from(i as u32 + 1 as u32);
+                let index_bn = BigInt::from(i as u32 + 1);
                 ECScalar::from(&index_bn)
             })
             .collect();
